add checks for non-image APODs/consolidate code
